pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ps-kinematics"
version = "0.1.0"
description = "Kinematic feature extraction from videos of the hand pronation-supination motor task for Parkinson's disease and PSP research."
readme = "README.md"
requires-python = ">=3.10"
license = { file = "LICENSE" }
authors = [{ name = "Mohamad Naseb Aljalab" }]
keywords = [
    "parkinson",
    "psp",
    "kinematics",
    "pose-estimation",
    "medical-imaging",
    "mds-updrs",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Medical Science Apps.",
    "Topic :: Scientific/Engineering :: Image Recognition",
]

# Core runtime dependencies required for the baseline MediaPipe pipeline.
# Heavier backends live in optional extras below.
dependencies = [
    "numpy>=1.26.4",
    "opencv-python==4.12.0.88",
    "pandas==2.3.3",
    "mediapipe==0.10.31",
    "scipy",
    "matplotlib==3.10.8",
    "tqdm==4.67.1",
    "psutil==7.2.1",
    "scikit-learn==1.8.0",
    "pyyaml",
    "setuptools<81",
]

[project.optional-dependencies]
refinement = [
    "ultralytics",
    "mmpose==1.3.2",
    "mmdet>=3.0",
    "mmengine",
    "mmcv>=2.0",
    "realesrgan",
    "basicsr",
]
ml = [
    "torch==2.1.2",
    "lightgbm",
    "optuna",
    "statsmodels",
    "factor-analyzer",
    "statannotations",
]
plots = [
    "seaborn==0.13.2",
    "openpyxl",
]
dev = [
    "ruff",
    "black",
]

[project.urls]
Repository = "https://github.com/mn-j/video-ps-kinematics"

[tool.setuptools.packages.find]
where = ["src"]
include = ["ps_kinematics*"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
ignore = [
    "E402",  # module-level import not at top — unavoidable with try/except availability guards
    "E501",  # line too long — handled by black
    "E731",  # lambda assignment — acceptable in callback patterns
    "E741",  # ambiguous variable name (l, O) — common in scientific code
    "F841",  # local variable assigned but unused — low priority, pre-existing
]

[tool.black]
line-length = 100
target-version = ["py310", "py311"]
